My 207 Project work
I have implemented the Round Robin and Weighted Round Robin algorithms to the POX SDN ip_loadbalancer. When the weights are given along with server IP's using the command line as 10.0.0.1:2,10.0.0.2:3,.... , the code automatically works in the WRR mode. If no weights are given, the Round Robin method will be called. 
This is my part of work in my CMPE 207 project.